Cleanup transform classes
